+++ client/src/pages/CandidateProgress.tsx
@@ -1,3 +1,0 @@
-export default function CandidateProgress(): JSX.Element {
-  return <div>CandidateProgress</div>;
-}