+++ client/src/pages/group_auth/utils/constants.ts
@@ -1,0 +1,11 @@
+import { HREF } from "../../../config/href"
+
+//  remarks: form login
+
+export const ROLE_DASHBOARD: Record<string, string> = {
+  candidate: HREF.CANDIDATE_DASHBOARD,
+  grade_1_assistant: HREF.ASSISTANT_DASHBOARD,
+  grade_2_manager: HREF.MANAGER_DASHBOARD,
+  grade_3_executive: HREF.MANAGER_DASHBOARD,
+  sys_admin: HREF.MANAGER_DASHBOARD,
+};