+++ client/src/pages/group_auth/utils/types.ts
@@ -1,0 +1,12 @@
+//  remarks centralisation
+
+export type AUTH_TYPES = {
+  login: TLogin
+}
+
+//  remarks: types
+
+type TLogin = {
+    username: string,
+    _password: string
+  }