+++ client/src/styles/color.ts
@@ -1,6 +1,8 @@
 export const COLORS = {
   dark_teal: '#0f766e',
   light_gray: '#dddddd',
+  gray: "#999999",
+  dark_gray: "#333333",
   success_teal: '#0d9488',
   error_red: '#991b1b',
   button_yellow: '#eab308',