+++ docs/change-log.md
@@ -18,7 +18,18 @@ This file records the major version changes of the projects.
 
 <br/>
 
-## [1.0.5] - Jul 14, 2026  - In Progress
+## [1.0.5] - Jul 18, 2026  - In Progress
+
+<i>Objective: build authentication-related interface and link navigation </i>
+
+### Added
+
+- implemented user sign-up, log-in, access control, cookie and password reset for authentication at the server-side.
+- implemented login page and its navigation to corresponding role-based dashboard.
+
+<br/>
+
+## [1.0.5] - Jul 14, 2026
 
 <i>Objective: build basic authentication and implement id-specific sub-page of candidates, staff and department table </i>
 
@@ -28,8 +39,7 @@ This file records the major version changes of the projects.
 - implemented candidate details, education, experience, test score and preference section at profile candidate page.
 - provided new content for setup and maintenance guides at `developer-guide.md`.
 - provided new authentication workflow chart at `architecture.md`.
-- implemented user sign-up, log-in, access control, cookie and password reset for authentication at the server-side.
-- applied the profile format into profile staff, following with loading spinner and error page redirection.
+- applied the profile format into profile staff and sys_user, following with loading spinner and error page redirection.
 
 ### Edited